+++ app/router.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative 'controller'
+require_relative '../controller/controller'
 
 class Router
